+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to deser are documented here.
 
 - Added support for `Box<T>`.
 - Added newtype struct support for derive feature.
+- `Driver` is now called `DeserializeDriver`.
 
 ## 0.6.0